Split PaRSEC backend ttg.h to make it maintainable.
