+++ ttg/ttg/parsec/tt.h
@@ -1,0 +1,1918 @@
+#ifndef TTG_PARSEC_TT_H
+#define TTG_PARSEC_TT_H
+
+#include <parsec.h>
+
+#include "ttg/parsec/ttg_data_copy.h"
+#include "ttg/parsec/task.h"
+#include "ttg/parsec/vars.h"
+#include "ttg/parsec/comm_helper.h"
+#include "ttg/parsec/hooks.h"
+#include "ttg/base/tt.h"
+#include "ttg/util/meta.h"
+#include "ttg/terminal.h"
+#include "ttg/edge.h"
+#include "ttg/serialization/data_descriptor.h"
+
+namespace ttg_parsec {
+
+
+  namespace detail {
+
+    struct ParsecTTBase {
+     protected:
+      //  static std::map<int, ParsecBaseTT*> function_id_to_instance;
+      parsec_hash_table_t tasks_table;
+      parsec_task_class_t self;
+    };
+
+  }  // namespace detail
+
+  template <typename keyT, typename output_terminalsT, typename derivedT, typename input_valueTs>
+  class TT : public ttg::TTBase, detail::ParsecTTBase {
+   private:
+    /// preconditions
+    static_assert(ttg::meta::is_typelist_v<input_valueTs>,
+                  "The fourth template for ttg::TT must be a ttg::typelist containing the input types");
+    // create a virtual control input if the input list is empty, to be used in invoke()
+    using actual_input_tuple_type = std::conditional_t<!ttg::meta::typelist_is_empty_v<input_valueTs>,
+                                                       ttg::meta::typelist_to_tuple_t<input_valueTs>, std::tuple<void>>;
+    using input_tuple_type = ttg::meta::typelist_to_tuple_t<input_valueTs>;
+    static_assert(ttg::meta::is_tuple_v<output_terminalsT>,
+                  "Second template argument for ttg::TT must be std::tuple containing the output terminal types");
+    static_assert((ttg::meta::none_has_reference_v<input_valueTs>), "Input typelist cannot contain reference types");
+    static_assert(ttg::meta::is_none_Void_v<input_valueTs>, "ttg::Void is for internal use only, do not use it");
+
+    parsec_mempool_t mempools;
+
+    // check for a non-type member named have_cuda_op
+    template <typename T>
+    using have_cuda_op_non_type_t = decltype(&T::have_cuda_op);
+
+    bool alive = true;
+
+    static constexpr int numinedges = std::tuple_size_v<input_tuple_type>;     // number of input edges
+    static constexpr int numins = std::tuple_size_v<actual_input_tuple_type>;  // number of input arguments
+    static constexpr int numouts = std::tuple_size_v<output_terminalsT>;       // number of outputs
+    static constexpr int numflows = std::max(numins, numouts);                 // max number of flows
+
+    /// @return true if derivedT::have_cuda_op exists and is defined to true
+    static constexpr bool derived_has_cuda_op() {
+      if constexpr (ttg::meta::is_detected_v<have_cuda_op_non_type_t, derivedT>) {
+        return derivedT::have_cuda_op;
+      } else {
+        return false;
+      }
+    }
+
+   public:
+    using ttT = TT;
+    using key_type = keyT;
+    using input_terminals_type = ttg::detail::input_terminals_tuple_t<keyT, input_tuple_type>;
+    using input_args_type = actual_input_tuple_type;
+    using input_edges_type = ttg::detail::edges_tuple_t<keyT, ttg::meta::decayed_typelist_t<input_tuple_type>>;
+    // if have data inputs and (always last) control input, convert last input to Void to make logic easier
+    using input_values_full_tuple_type =
+        ttg::meta::void_to_Void_tuple_t<ttg::meta::decayed_typelist_t<actual_input_tuple_type>>;
+    using input_refs_full_tuple_type =
+        ttg::meta::add_glvalue_reference_tuple_t<ttg::meta::void_to_Void_tuple_t<actual_input_tuple_type>>;
+    using input_values_tuple_type = ttg::meta::drop_void_t<ttg::meta::decayed_typelist_t<input_tuple_type>>;
+    using input_refs_tuple_type = ttg::meta::drop_void_t<ttg::meta::add_glvalue_reference_tuple_t<input_tuple_type>>;
+
+    static constexpr int numinvals =
+        std::tuple_size_v<input_refs_tuple_type>;  // number of input arguments with values (i.e. omitting the control
+                                                   // input, if any)
+
+    using output_terminals_type = output_terminalsT;
+    using output_edges_type = typename ttg::terminals_to_edges<output_terminalsT>::type;
+
+    template <std::size_t i, typename resultT, typename InTuple>
+    static resultT get(InTuple &&intuple) {
+      return static_cast<resultT>(std::get<i>(std::forward<InTuple>(intuple)));
+    };
+    template <std::size_t i, typename InTuple>
+    static auto &get(InTuple &&intuple) {
+      return std::get<i>(std::forward<InTuple>(intuple));
+    };
+
+   private:
+    using task_t = detail::parsec_ttg_task_t<ttT>;
+
+    friend task_t;
+
+    /* the offset of the key placed after the task structure in the memory from mempool */
+    constexpr static const size_t task_key_offset = sizeof(task_t);
+
+    input_terminals_type input_terminals;
+    output_terminalsT output_terminals;
+
+   protected:
+    const auto &get_output_terminals() const { return output_terminals; }
+
+   private:
+    template <std::size_t... IS>
+    static constexpr auto make_set_args_fcts(std::index_sequence<IS...>) {
+      using resultT = decltype(set_arg_from_msg_fcts);
+      return resultT{{&TT::set_arg_from_msg<IS>...}};
+    }
+    constexpr static std::array<void (TT::*)(void *, std::size_t), numins> set_arg_from_msg_fcts =
+        make_set_args_fcts(std::make_index_sequence<numins>{});
+
+    template <std::size_t... IS>
+    static constexpr auto make_set_size_fcts(std::index_sequence<IS...>) {
+      using resultT = decltype(set_argstream_size_from_msg_fcts);
+      return resultT{{&TT::argstream_set_size_from_msg<IS>...}};
+    }
+    constexpr static std::array<void (TT::*)(void *, std::size_t), numins> set_argstream_size_from_msg_fcts =
+        make_set_size_fcts(std::make_index_sequence<numins>{});
+
+    template <std::size_t... IS>
+    static constexpr auto make_finalize_argstream_fcts(std::index_sequence<IS...>) {
+      using resultT = decltype(finalize_argstream_from_msg_fcts);
+      return resultT{{&TT::finalize_argstream_from_msg<IS>...}};
+    }
+    constexpr static std::array<void (TT::*)(void *, std::size_t), numins> finalize_argstream_from_msg_fcts =
+        make_finalize_argstream_fcts(std::make_index_sequence<numins>{});
+
+    template <std::size_t... IS>
+    static constexpr auto make_get_from_pull_fcts(std::index_sequence<IS...>) {
+      using resultT = decltype(get_from_pull_msg_fcts);
+      return resultT{{&TT::get_from_pull_msg<IS>...}};
+    }
+    constexpr static std::array<void (TT::*)(void *, std::size_t), numinedges> get_from_pull_msg_fcts =
+        make_get_from_pull_fcts(std::make_index_sequence<numinedges>{});
+
+    ttg::World world;
+    ttg::meta::detail::keymap_t<keyT> keymap;
+    ttg::meta::detail::keymap_t<keyT> priomap;
+    // For now use same type for unary/streaming input terminals, and stream reducers assigned at runtime
+    ttg::meta::detail::input_reducers_t<actual_input_tuple_type>
+        input_reducers;  //!< Reducers for the input terminals (empty = expect single value)
+    std::array<std::size_t, numins> static_stream_goal;
+    int num_pullins = 0;
+
+   public:
+    ttg::World get_world() const { return world; }
+
+   private:
+    /// dispatches a call to derivedT::op if Space == Host, otherwise to derivedT::op_cuda if Space == CUDA
+    template <ttg::ExecutionSpace Space, typename... Args>
+    void op(Args &&...args) {
+      derivedT *derived = static_cast<derivedT *>(this);
+      if constexpr (Space == ttg::ExecutionSpace::Host)
+        derived->op(std::forward<Args>(args)...);
+      else if constexpr (Space == ttg::ExecutionSpace::CUDA)
+        derived->op_cuda(std::forward<Args>(args)...);
+      else
+        abort();
+    }
+
+    template <std::size_t i, typename terminalT, typename Key>
+    void invoke_pull_terminal(terminalT &in, const Key &key, detail::parsec_ttg_task_base_t *task) {
+      if (in.is_pull_terminal) {
+        auto owner = in.container.owner(key);
+        if (owner != world.rank()) {
+          get_pull_terminal_data_from<i>(owner, key);
+        } else {
+          // push the data to the task
+          set_arg<i>(key, (in.container).get(key));
+        }
+      }
+    }
+
+    template <std::size_t i, typename Key>
+    void get_pull_terminal_data_from(const int owner,
+                                     const Key &key) {
+      using msg_t = detail::msg_t;
+      auto &world_impl = world.impl();
+      parsec_taskpool_t *tp = world_impl.taskpool();
+      std::unique_ptr<msg_t> msg = std::make_unique<msg_t>(get_instance_id(), tp->taskpool_id,
+                                                           detail::msg_header_t::MSG_GET_FROM_PULL, i, 1);
+      /* pack the key */
+      size_t pos = 0;
+      pos = pack(key, msg->bytes, pos);
+      tp->tdm.module->outgoing_message_start(tp, owner, NULL);
+      tp->tdm.module->outgoing_message_pack(tp, owner, NULL, NULL, 0);
+      parsec_ce.send_am(&parsec_ce, world_impl.parsec_ttg_tag(), owner, static_cast<void *>(msg.get()),
+                        sizeof(detail::msg_header_t) + pos);
+    }
+
+    template <std::size_t... IS, typename Key = keyT>
+    void invoke_pull_terminals(std::index_sequence<IS...>, const Key &key, detail::parsec_ttg_task_base_t *task) {
+      int junk[] = {0, (invoke_pull_terminal<IS>(
+                            std::get<IS>(input_terminals), key, task),
+                        0)...};
+      junk[0]++;
+    }
+
+    template <std::size_t... IS>
+    static input_refs_tuple_type make_tuple_of_ref_from_array(task_t *task, std::index_sequence<IS...>) {
+      return input_refs_tuple_type{static_cast<std::tuple_element_t<IS, input_refs_tuple_type>>(
+          *reinterpret_cast<std::remove_reference_t<std::tuple_element_t<IS, input_refs_tuple_type>> *>(
+              task->parsec_task.data[IS].data_in->device_private))...};
+    }
+
+    template <ttg::ExecutionSpace Space>
+    static void static_op(parsec_task_t *parsec_task) {
+      task_t *task = (task_t*)parsec_task;
+      ttT *baseobj = task->tt;
+      derivedT *obj = static_cast<derivedT *>(baseobj);
+      assert(detail::parsec_ttg_caller == NULL);
+      detail::parsec_ttg_caller = static_cast<detail::parsec_ttg_task_base_t*>(task);
+      if (obj->tracing()) {
+        if constexpr (!ttg::meta::is_void_v<keyT>)
+          ttg::trace(obj->get_world().rank(), ":", obj->get_name(), " : ", task->key, ": executing");
+        else
+          ttg::trace(obj->get_world().rank(), ":", obj->get_name(), " : executing");
+      }
+
+      if constexpr (!ttg::meta::is_void_v<keyT> && !ttg::meta::is_empty_tuple_v<input_values_tuple_type>) {
+        auto input = make_tuple_of_ref_from_array(task, std::make_index_sequence<numinvals>{});
+        baseobj->template op<Space>(task->key, std::move(input), obj->output_terminals);
+      } else if constexpr (!ttg::meta::is_void_v<keyT> && ttg::meta::is_empty_tuple_v<input_values_tuple_type>) {
+        baseobj->template op<Space>(task->key, obj->output_terminals);
+      } else if constexpr (ttg::meta::is_void_v<keyT> && !ttg::meta::is_empty_tuple_v<input_values_tuple_type>) {
+        auto input = make_tuple_of_ref_from_array(task, std::make_index_sequence<numinvals>{});
+        baseobj->template op<Space>(std::move(input), obj->output_terminals);
+      } else if constexpr (ttg::meta::is_void_v<keyT> && ttg::meta::is_empty_tuple_v<input_values_tuple_type>) {
+        baseobj->template op<Space>(obj->output_terminals);
+      } else {
+        abort();
+      }
+      detail::parsec_ttg_caller = NULL;
+
+      if (obj->tracing()) {
+        if constexpr (!ttg::meta::is_void_v<keyT>)
+          ttg::trace(obj->get_world().rank(), ":", obj->get_name(), " : ", task->key, ": done executing");
+        else
+          ttg::trace(obj->get_world().rank(), ":", obj->get_name(), " : done executing");
+      }
+    }
+
+    template <ttg::ExecutionSpace Space>
+    static void static_op_noarg(parsec_task_t *parsec_task) {
+      task_t *task = static_cast<task_t*>(parsec_task);
+      ttT *baseobj = (ttT *)task->object_ptr;
+      derivedT *obj = (derivedT *)task->object_ptr;
+      assert(detail::parsec_ttg_caller == NULL);
+      detail::parsec_ttg_caller = task;
+      if constexpr (!ttg::meta::is_void_v<keyT>) {
+        baseobj->template op<Space>(task->key, obj->output_terminals);
+      } else if constexpr (ttg::meta::is_void_v<keyT>) {
+        baseobj->template op<Space>(obj->output_terminals);
+      } else
+        abort();
+      detail::parsec_ttg_caller = NULL;
+    }
+
+   protected:
+    template <typename T>
+    uint64_t unpack(T &obj, void *_bytes, uint64_t pos) {
+      const ttg_data_descriptor *dObj = ttg::get_data_descriptor<ttg::meta::remove_cvr_t<T>>();
+      uint64_t payload_size;
+      if constexpr (!ttg::default_data_descriptor<ttg::meta::remove_cvr_t<T>>::serialize_size_is_const) {
+        const ttg_data_descriptor *dSiz = ttg::get_data_descriptor<uint64_t>();
+        dSiz->unpack_payload(&payload_size, sizeof(uint64_t), pos, _bytes);
+        pos += sizeof(uint64_t);
+      } else {
+        payload_size = dObj->payload_size(&obj);
+      }
+      dObj->unpack_payload(&obj, payload_size, pos, _bytes);
+      return pos + payload_size;
+    }
+
+    template <typename T>
+    uint64_t pack(T &obj, void *bytes, uint64_t pos) {
+      const ttg_data_descriptor *dObj = ttg::get_data_descriptor<ttg::meta::remove_cvr_t<T>>();
+      uint64_t payload_size = dObj->payload_size(&obj);
+      if constexpr (!ttg::default_data_descriptor<ttg::meta::remove_cvr_t<T>>::serialize_size_is_const) {
+        const ttg_data_descriptor *dSiz = ttg::get_data_descriptor<uint64_t>();
+        dSiz->pack_payload(&payload_size, sizeof(uint64_t), pos, bytes);
+        pos += sizeof(uint64_t);
+      }
+      dObj->pack_payload(&obj, payload_size, pos, bytes);
+      return pos + payload_size;
+    }
+
+    static void static_set_arg(void *data, std::size_t size, ttg::TTBase *bop) {
+      assert(size >= sizeof(detail::msg_header_t) &&
+             "Trying to unpack as message that does not hold enough bytes to represent a single header");
+      detail::msg_header_t *hd = static_cast<detail::msg_header_t *>(data);
+      derivedT *obj = reinterpret_cast<derivedT *>(bop);
+      switch (hd->fn_id) {
+        case detail::msg_header_t::MSG_SET_ARG: {
+          if (0 <= hd->param_id) {
+            assert(hd->param_id >= 0);
+            assert(hd->param_id < obj->set_arg_from_msg_fcts.size());
+            auto member = obj->set_arg_from_msg_fcts[hd->param_id];
+            (obj->*member)(data, size);
+          } else {
+            // there is no good reason to have negative param ids
+            abort();
+          }
+          break;
+        }
+        case detail::msg_header_t::MSG_SET_ARGSTREAM_SIZE: {
+          assert(hd->param_id >= 0);
+          assert(hd->param_id < obj->set_argstream_size_from_msg_fcts.size());
+          auto member = obj->set_argstream_size_from_msg_fcts[hd->param_id];
+          (obj->*member)(data, size);
+          break;
+        }
+        case detail::msg_header_t::MSG_FINALIZE_ARGSTREAM_SIZE: {
+          assert(hd->param_id >= 0);
+          assert(hd->param_id < obj->finalize_argstream_from_msg_fcts.size());
+          auto member = obj->finalize_argstream_from_msg_fcts[hd->param_id];
+          (obj->*member)(data, size);
+          break;
+        }
+        case detail::msg_header_t::MSG_GET_FROM_PULL: {
+          assert(hd->param_id >= 0);
+          assert(hd->param_id < obj->get_from_pull_msg_fcts.size());
+          auto member = obj->get_from_pull_msg_fcts[hd->param_id];
+          (obj->*member)(data, size);
+          break;
+        }
+        default:
+          abort();
+      }
+    }
+
+    /** Returns the task memory pool owned by the calling thread */
+    inline parsec_thread_mempool_t *get_task_mempool(void) {
+      auto &world_impl = world.impl();
+      parsec_execution_stream_s *es = world_impl.execution_stream();
+      int index = (es->virtual_process->vp_id * es->virtual_process->nb_cores + es->th_id);
+      return &mempools.thread_mempools[index];
+    }
+
+    template <size_t i, typename valueT>
+    void set_arg_from_msg_keylist(ttg::span<keyT> &&keylist, detail::ttg_data_copy_t *copy) {
+      /* create a dummy task that holds the copy, which can be reused by others */
+      task_t *dummy;
+      parsec_execution_stream_s *es = world.impl().execution_stream();
+      parsec_thread_mempool_t *mempool = get_task_mempool();
+      dummy = new (parsec_thread_mempool_allocate(mempool)) task_t(mempool, &this->self);
+      // TODO: do we need to copy static_stream_goal in dummy?
+
+      /* set the received value as the dummy's only data */
+      dummy->parsec_task.data[0].data_in = copy;
+
+      /* save the current task and set the dummy task */
+      auto parsec_ttg_caller_save = detail::parsec_ttg_caller;
+      detail::parsec_ttg_caller = dummy;
+
+      /* iterate over the keys and have them use the copy we made */
+      parsec_task_t *task_ring = nullptr;
+      for (auto &&key : keylist) {
+        set_arg_local_impl<i>(key, *reinterpret_cast<valueT *>(copy->device_private), copy, &task_ring);
+      }
+
+      if (nullptr != task_ring) {
+        auto &world_impl = world.impl();
+        __parsec_schedule(world_impl.execution_stream(), task_ring, 0);
+      }
+
+      /* restore the previous task */
+      detail::parsec_ttg_caller = parsec_ttg_caller_save;
+
+      /* release the dummy task */
+      complete_task_and_release(es, &dummy->parsec_task);
+      parsec_thread_mempool_free(mempool, &dummy->parsec_task);
+    }
+
+    // there are 6 types of set_arg:
+    // - case 1: nonvoid Key, complete Value type
+    // - case 2: nonvoid Key, void Value, mixed (data+control) inputs
+    // - case 3: nonvoid Key, void Value, no inputs
+    // - case 4:    void Key, complete Value type
+    // - case 5:    void Key, void Value, mixed (data+control) inputs
+    // - case 6:    void Key, void Value, no inputs
+    // implementation of these will be further split into "local-only" and global+local
+
+    template <std::size_t i>
+    void set_arg_from_msg(void *data, std::size_t size) {
+      using valueT = std::tuple_element_t<i, actual_input_tuple_type>;
+      using msg_t = detail::msg_t;
+      msg_t *msg = static_cast<msg_t *>(data);
+      if constexpr (!ttg::meta::is_void_v<keyT>) {
+        /* unpack the keys */
+        uint64_t pos = 0;
+        std::vector<keyT> keylist;
+        int num_keys = msg->tt_id.num_keys;
+        keylist.reserve(num_keys);
+        auto rank = world.rank();
+        for (int k = 0; k < num_keys; ++k) {
+          keyT key;
+          pos = unpack(key, msg->bytes, pos);
+          assert(keymap(key) == rank);
+          keylist.push_back(std::move(key));
+        }
+        // case 1
+        if constexpr (!ttg::meta::is_void_v<valueT>) {
+          using decvalueT = std::decay_t<valueT>;
+          if constexpr (!ttg::has_split_metadata<decvalueT>::value) {
+            detail::ttg_data_copy_t *copy = detail::create_new_datacopy(decvalueT{});
+            unpack(*static_cast<decvalueT *>(copy->device_private), msg->bytes, pos);
+
+            set_arg_from_msg_keylist<i, decvalueT>(ttg::span<keyT>(&keylist[0], num_keys), copy);
+          } else {
+            /* unpack the header and start the RMA transfers */
+            ttg::SplitMetadataDescriptor<decvalueT> descr;
+            using metadata_t = decltype(descr.get_metadata(std::declval<decvalueT>()));
+            size_t metadata_size = sizeof(metadata_t);
+
+            /* unpack the metadata */
+            metadata_t metadata;
+            std::memcpy(&metadata, msg->bytes + pos, metadata_size);
+            pos += metadata_size;
+
+            /* unpack the remote rank */
+            int remote;
+            std::memcpy(&remote, msg->bytes + pos, sizeof(remote));
+            pos += sizeof(remote);
+
+            assert(remote < world.size());
+
+            /* extract the number of chunks */
+            int32_t num_iovecs;
+            std::memcpy(&num_iovecs, msg->bytes + pos, sizeof(num_iovecs));
+            pos += sizeof(num_iovecs);
+
+            detail::ttg_data_copy_t *copy = detail::create_new_datacopy(descr.create_from_metadata(metadata));
+            /* nothing else to do if the object is empty */
+            if (0 == num_iovecs) {
+              set_arg_from_msg_keylist<i, decvalueT>(keylist, copy);
+            } else {
+              /* extract the callback tag */
+              parsec_ce_tag_t cbtag;
+              std::memcpy(&cbtag, msg->bytes + pos, sizeof(cbtag));
+              pos += sizeof(cbtag);
+
+              /* create the value from the metadata */
+              auto activation = new detail::rma_delayed_activate(
+                  std::move(keylist), copy, num_iovecs, [this](std::vector<keyT> &&keylist, detail::ttg_data_copy_t *copy) {
+                    set_arg_from_msg_keylist<i, decvalueT>(keylist, copy);
+                    this->world.impl().decrement_inflight_msg();
+                  });
+              auto &val = *static_cast<decvalueT *>(copy->device_private);
+
+              using ActivationT = std::decay_t<decltype(*activation)>;
+
+              int nv = 0;
+              /* process payload iovecs */
+              auto iovecs = descr.get_data(val);
+              /* start the RMA transfers */
+              for (auto &&iov : iovecs) {
+                ++nv;
+                parsec_ce_mem_reg_handle_t rreg;
+                int32_t rreg_size_i;
+                std::memcpy(&rreg_size_i, msg->bytes + pos, sizeof(rreg_size_i));
+                pos += sizeof(rreg_size_i);
+                rreg = static_cast<parsec_ce_mem_reg_handle_t>(msg->bytes + pos);
+                pos += rreg_size_i;
+                // std::intptr_t *fn_ptr = reinterpret_cast<std::intptr_t *>(msg->bytes + pos);
+                // pos += sizeof(*fn_ptr);
+                std::intptr_t fn_ptr;
+                std::memcpy(&fn_ptr, msg->bytes + pos, sizeof(fn_ptr));
+                pos += sizeof(fn_ptr);
+
+                /* register the local memory */
+                parsec_ce_mem_reg_handle_t lreg;
+                size_t lreg_size;
+                parsec_ce.mem_register(iov.data, PARSEC_MEM_TYPE_NONCONTIGUOUS, iov.num_bytes, parsec_datatype_int8_t,
+                                       iov.num_bytes, &lreg, &lreg_size);
+                world.impl().increment_inflight_msg();
+                /* TODO: PaRSEC should treat the remote callback as a tag, not a function pointer! */
+                parsec_ce.get(&parsec_ce, lreg, 0, rreg, 0, iov.num_bytes, remote,
+                              &detail::get_complete_cb<ActivationT>, activation,
+                              /*world.impl().parsec_ttg_rma_tag()*/
+                              cbtag, &fn_ptr, sizeof(std::intptr_t));
+              }
+
+              assert(num_iovecs == nv);
+              assert(size == (pos + sizeof(detail::msg_header_t)));
+            }
+          }
+          // case 2 and 3
+        } else if constexpr (!ttg::meta::is_void_v<keyT> && std::is_void_v<valueT>) {
+          for (auto &&key : keylist) {
+            set_arg<i, keyT, ttg::Void>(key, ttg::Void{});
+          }
+        }
+        // case 4
+      } else if constexpr (ttg::meta::is_void_v<keyT> && !std::is_void_v<valueT>) {
+        using decvalueT = std::decay_t<valueT>;
+        decvalueT val;
+        /* TODO: handle split-metadata case as with non-void keys */
+        unpack(val, msg->bytes, 0);
+        set_arg<i, keyT, valueT>(std::move(val));
+        // case 5 and 6
+      } else if constexpr (ttg::meta::is_void_v<keyT> && std::is_void_v<valueT>) {
+        set_arg<i, keyT, ttg::Void>(ttg::Void{});
+      } else {
+        abort();
+      }
+    }
+
+    template <std::size_t i>
+    void finalize_argstream_from_msg(void *data, std::size_t size) {
+      using msg_t = detail::msg_t;
+      msg_t *msg = static_cast<msg_t *>(data);
+      if constexpr (!ttg::meta::is_void_v<keyT>) {
+        /* unpack the key */
+        uint64_t pos = 0;
+        auto rank = world.rank();
+        keyT key;
+        pos = unpack(key, msg->bytes, pos);
+        assert(keymap(key) == rank);
+        finalize_argstream<i>(key);
+      } else {
+        auto rank = world.rank();
+        assert(keymap() == rank);
+        finalize_argstream<i>();
+      }
+    }
+
+    template <std::size_t i>
+    void argstream_set_size_from_msg(void *data, std::size_t size) {
+      using msg_t = detail::msg_t;
+      auto msg = static_cast<msg_t *>(data);
+      uint64_t pos = 0;
+      if constexpr (!ttg::meta::is_void_v<keyT>) {
+        /* unpack the key */
+        auto rank = world.rank();
+        keyT key;
+        pos = unpack(key, msg->bytes, pos);
+        assert(keymap(key) == rank);
+        std::size_t argstream_size;
+        pos = unpack(argstream_size, msg->bytes, pos);
+        set_argstream_size<i>(key, argstream_size);
+      } else {
+        auto rank = world.rank();
+        assert(keymap() == rank);
+        std::size_t argstream_size;
+        pos = unpack(argstream_size, msg->bytes, pos);
+        set_argstream_size<i>(argstream_size);
+      }
+    }
+
+    template <std::size_t i>
+    void get_from_pull_msg(void *data, std::size_t size) {
+      using msg_t = detail::msg_t;
+      msg_t *msg = static_cast<msg_t *>(data);
+      auto &in = std::get<i>(input_terminals);
+      if constexpr (!ttg::meta::is_void_v<keyT>) {
+        /* unpack the key */
+        uint64_t pos = 0;
+        keyT key;
+        pos = unpack(key, msg->bytes, pos);
+        set_arg<i>(key, (in.container).get(key));
+      }
+    }
+
+    template <std::size_t i, typename Key, typename Value>
+    std::enable_if_t<!ttg::meta::is_void_v<Key> && !std::is_void_v<std::decay_t<Value>>, void> set_arg_local(
+        const Key &key, Value &&value) {
+      set_arg_local_impl<i>(key, std::forward<Value>(value));
+    }
+
+    template <std::size_t i, typename Key = keyT, typename Value>
+    std::enable_if_t<ttg::meta::is_void_v<Key> && !std::is_void_v<std::decay_t<Value>>, void> set_arg_local(
+        Value &&value) {
+      set_arg_local_impl<i>(ttg::Void{}, std::forward<Value>(value));
+    }
+
+    template <std::size_t i, typename Key, typename Value>
+    std::enable_if_t<!ttg::meta::is_void_v<Key> && !std::is_void_v<std::decay_t<Value>>, void> set_arg_local(
+        const Key &key, const Value &value) {
+      set_arg_local_impl<i>(key, value);
+    }
+
+    template <std::size_t i, typename Key = keyT, typename Value>
+    std::enable_if_t<ttg::meta::is_void_v<Key> && !std::is_void_v<std::decay_t<Value>>, void> set_arg_local(
+        const Value &value) {
+      set_arg_local_impl<i>(ttg::Void{}, value);
+    }
+
+    template <std::size_t i, typename Key = keyT, typename Value>
+    std::enable_if_t<ttg::meta::is_void_v<Key> && !std::is_void_v<std::decay_t<Value>>, void> set_arg_local(
+        std::shared_ptr<const Value> &valueptr) {
+      set_arg_local_impl<i>(ttg::Void{}, *valueptr);
+    }
+
+    template <typename Key>
+    task_t *create_new_task(const Key &key) {
+      constexpr const bool keyT_is_Void = ttg::meta::is_void_v<keyT>;
+      auto &world_impl = world.impl();
+      task_t *newtask;
+      parsec_thread_mempool_t *mempool = get_task_mempool();
+      char *taskobj = (char *)parsec_thread_mempool_allocate(mempool);
+      int32_t priority;
+      if constexpr (!keyT_is_Void) {
+        priority = priomap(key);
+        /* placement-new the task */
+        newtask = new (taskobj) task_t(key, mempool, &this->self, world_impl.taskpool(), this, priority);
+      } else {
+        priority = priomap();
+        /* placement-new the task */
+        newtask = new (taskobj) task_t(mempool, &this->self, world_impl.taskpool(), this, priority);
+      }
+
+      newtask->function_template_class_ptr[static_cast<std::size_t>(ttg::ExecutionSpace::Host)] =
+          reinterpret_cast<detail::parsec_static_op_t>(&TT::static_op<ttg::ExecutionSpace::Host>);
+      if constexpr (derived_has_cuda_op())
+        newtask->function_template_class_ptr[static_cast<std::size_t>(ttg::ExecutionSpace::CUDA)] =
+            reinterpret_cast<detail::parsec_static_op_t>(&TT::static_op<ttg::ExecutionSpace::CUDA>);
+
+      for (int i = 0; i < static_stream_goal.size(); ++i) {
+        newtask->stream[i].goal = static_stream_goal[i];
+      }
+
+      ttg::trace(world.rank(), ":", get_name(), " : ", key, ": creating task");
+      return newtask;
+    }
+
+    // Used to set the i'th argument
+    template <std::size_t i, typename Key, typename Value>
+    void set_arg_local_impl(const Key &key, Value &&value, detail::ttg_data_copy_t *copy_in = nullptr,
+                            parsec_task_t **task_ring = nullptr) {
+      using valueT = std::tuple_element_t<i, input_values_full_tuple_type>;
+      constexpr const bool input_is_const = std::is_const_v<std::tuple_element_t<i, input_args_type>>;
+      constexpr const bool valueT_is_Void = ttg::meta::is_void_v<valueT>;
+      constexpr const bool keyT_is_Void = ttg::meta::is_void_v<Key>;
+
+      if constexpr (!valueT_is_Void) {
+        ttg::trace(world.rank(), ":", get_name(), " : ", key, ": received value for argument : ", i,
+                   " : value = ", value);
+      } else {
+        ttg::trace(world.rank(), ":", get_name(), " : ", key, ": received value for argument : ", i);
+      }
+
+      parsec_key_t hk = 0;
+      if constexpr (!keyT_is_Void) {
+        hk = reinterpret_cast<parsec_key_t>(&key);
+        assert(keymap(key) == world.rank());
+      }
+
+      task_t *task;
+      auto &world_impl = world.impl();
+      auto &reducer = std::get<i>(input_reducers);
+      bool release = true;
+      bool remove_from_hash = true;
+      bool get_pull_data = false;
+      /* If we have only one input and no reducer on that input we can skip the hash table */
+      if (numins > 1 || reducer) {
+        parsec_hash_table_lock_bucket(&tasks_table, hk);
+        if (nullptr == (task = (task_t *)parsec_hash_table_nolock_find(&tasks_table, hk))) {
+          task = create_new_task(key);
+          world_impl.increment_created();
+          parsec_hash_table_nolock_insert(&tasks_table, &task->tt_ht_item);
+          get_pull_data = !is_lazy_pull();
+        } else if (!reducer && numins == (task->in_data_count+1)) {
+          /* remove while we have the lock */
+          parsec_hash_table_nolock_remove(&tasks_table, hk);
+          remove_from_hash = false;
+        }
+        parsec_hash_table_unlock_bucket(&tasks_table, hk);
+      } else {
+        task = create_new_task(key);
+        world_impl.increment_created();
+        remove_from_hash = false;
+      }
+
+      if (reducer) {  // is this a streaming input? reduce the received value
+        // N.B. Right now reductions are done eagerly, without spawning tasks
+        //      this means we must lock
+        parsec_hash_table_lock_bucket(&tasks_table, hk);
+
+        if constexpr (!ttg::meta::is_void_v<valueT>) {  // for data values
+          // have a value already? if not, set, otherwise reduce
+          detail::ttg_data_copy_t *copy = nullptr;
+          if (nullptr == (copy = static_cast<detail::ttg_data_copy_t *>(task->parsec_task.data[i].data_in))) {
+            using decay_valueT = std::decay_t<valueT>;
+            /* For now, we always create a copy because we cannot rely on the task_release
+             * mechanism (it would release the task, not the reduction value). */
+            copy = detail::create_new_datacopy(std::forward<Value>(value));
+            task->parsec_task.data[i].data_in = copy;
+          } else {
+            reducer(*reinterpret_cast<std::decay_t<valueT> *>(copy->device_private), value);
+          }
+        } else {
+          reducer();  // even if this was a control input, must execute the reducer for possible side effects
+        }
+        task->stream[i].size++;
+        release = (task->stream[i].size == task->stream[i].goal);
+        if (release) {
+          parsec_hash_table_nolock_remove(&tasks_table, hk);
+          remove_from_hash = false;
+        }
+        parsec_hash_table_unlock_bucket(&tasks_table, hk);
+      } else {
+        /* whether the task needs to be deferred or not */
+        if constexpr (!valueT_is_Void) {
+          if (nullptr != task->parsec_task.data[i].data_in) {
+            ttg::print_error(get_name(), " : ", key, ": error argument is already set : ", i);
+            throw std::logic_error("bad set arg");
+          }
+
+          detail::ttg_data_copy_t *copy = copy_in;
+          if (nullptr == copy_in && nullptr != detail::parsec_ttg_caller) {
+            copy = detail::find_copy_in_task(detail::parsec_ttg_caller, &value);
+          }
+
+          if (nullptr != copy) {
+            /* register_data_copy might provide us with a different copy if !input_is_const */
+            copy = detail::register_data_copy<valueT>(copy, task, input_is_const);
+          } else {
+            copy = detail::create_new_datacopy(std::forward<Value>(value));
+          }
+          /* if we registered as a writer and were the first to register with this copy
+           * we need to defer the release of this task to give other tasks a chance to
+           * make a copy of the original data */
+          release = (copy->push_task == nullptr);
+          task->parsec_task.data[i].data_in = copy;
+        }
+      }
+      task->remove_from_hash = remove_from_hash;
+      if (release) {
+        release_task(task, task_ring);
+      }
+      /* if not pulling lazily, pull the data here */
+      if constexpr (!ttg::meta::is_void_v<keyT>) {
+        if (get_pull_data) {
+          invoke_pull_terminals(std::make_index_sequence<std::tuple_size_v<input_values_tuple_type>>{}, task->key, task);
+        }
+      }
+    }
+
+    void release_task(task_t *task,
+                      parsec_task_t **task_ring = nullptr) {
+      constexpr const bool keyT_is_Void = ttg::meta::is_void_v<keyT>;
+
+      /* if remove_from_hash == false, someone has already removed the task from the hash table
+       * so we know that the task is ready, no need to do atomic increments here */
+      bool is_ready = !task->remove_from_hash;
+      int32_t count;
+      if (is_ready) {
+        count = numins;
+      } else {
+        count = parsec_atomic_fetch_inc_int32(&task->in_data_count) + 1;
+        assert(count <= self.dependencies_goal);
+      }
+
+      auto &world_impl = world.impl();
+      ttT *baseobj = task->tt;
+
+      if (count == numins) {
+        parsec_execution_stream_t *es = world_impl.execution_stream();
+        parsec_key_t hk = task->pkey();
+        if (tracing()) {
+          if constexpr (!keyT_is_Void) {
+            ttg::trace(world.rank(), ":", get_name(), " : ", task->key, ": submitting task for op ");
+          } else {
+            ttg::trace(world.rank(), ":", get_name(), ": submitting task for op ");
+          }
+        }
+        if (task->remove_from_hash) parsec_hash_table_remove(&tasks_table, hk);
+        if (nullptr == task_ring) {
+          __parsec_schedule(es, &task->parsec_task, 0);
+        } else if (*task_ring == nullptr) {
+          /* the first task is set directly */
+          *task_ring = &task->parsec_task;
+        } else {
+          /* push into the ring */
+          parsec_list_item_ring_push_sorted(&(*task_ring)->super, &task->parsec_task.super,
+                                            offsetof(parsec_task_t, priority));
+        }
+      } else if constexpr (!ttg::meta::is_void_v<keyT>) {
+        if ((baseobj->num_pullins + count == numins) && baseobj->is_lazy_pull()) {
+          /* lazily pull the pull terminal data */
+          baseobj->invoke_pull_terminals(std::make_index_sequence<std::tuple_size_v<input_values_tuple_type>>{}, task->key, task);
+        }
+      }
+    }
+
+    // cases 1+2
+    template <std::size_t i, typename Key, typename Value>
+    std::enable_if_t<!ttg::meta::is_void_v<Key> && !std::is_void_v<std::decay_t<Value>>, void> set_arg(const Key &key,
+                                                                                                       Value &&value) {
+      set_arg_impl<i>(key, std::forward<Value>(value));
+    }
+
+    // cases 4+5+6
+    template <std::size_t i, typename Key, typename Value>
+    std::enable_if_t<ttg::meta::is_void_v<Key> && !std::is_void_v<std::decay_t<Value>>, void> set_arg(Value &&value) {
+      set_arg_impl<i>(ttg::Void{}, std::forward<Value>(value));
+    }
+
+    template <std::size_t i, typename Key = keyT>
+    std::enable_if_t<ttg::meta::is_void_v<Key>, void> set_arg() {
+      set_arg_impl<i>(ttg::Void{}, ttg::Void{});
+    }
+
+    // case 3
+    template <std::size_t i, typename Key>
+    std::enable_if_t<!ttg::meta::is_void_v<Key>, void> set_arg(const Key &key) {
+      set_arg_impl<i>(key, ttg::Void{});
+    }
+
+    // Used to set the i'th argument
+    template <std::size_t i, typename Key, typename Value>
+    void set_arg_impl(const Key &key, Value &&value) {
+      int owner;
+
+      if constexpr (!ttg::meta::is_void_v<Key>)
+        owner = keymap(key);
+      else
+        owner = keymap();
+      if (owner == world.rank()) {
+        if constexpr (!ttg::meta::is_void_v<keyT>)
+          set_arg_local<i, keyT, Value>(key, std::forward<Value>(value));
+        else
+          set_arg_local<i, keyT, Value>(std::forward<Value>(value));
+        return;
+      }
+      // the target task is remote. Pack the information and send it to
+      // the corresponding peer.
+      // TODO do we need to copy value?
+      using msg_t = detail::msg_t;
+      auto &world_impl = world.impl();
+      uint64_t pos = 0;
+      std::unique_ptr<msg_t> msg = std::make_unique<msg_t>(get_instance_id(), world_impl.taskpool()->taskpool_id,
+                                                           detail::msg_header_t::MSG_SET_ARG, i, 1);
+      using decvalueT = std::decay_t<Value>;
+      /* pack the key */
+      msg->tt_id.num_keys = 0;
+      if constexpr (!ttg::meta::is_void_v<Key>) {
+        pos = pack(key, msg->bytes, pos);
+        msg->tt_id.num_keys = 1;
+      }
+
+      if constexpr (!ttg::meta::is_void_v<decvalueT>) {
+        if constexpr (!ttg::has_split_metadata<decvalueT>::value) {
+          pos = pack(value, msg->bytes, pos);
+        } else {
+          detail::ttg_data_copy_t *copy;
+          copy = detail::find_copy_in_task(detail::parsec_ttg_caller, &value);
+          if (nullptr == copy) {
+            // We need to create a copy for this data, as it does not exist yet.
+            copy = detail::create_new_datacopy(std::forward<Value>(value));
+          }
+          copy = detail::register_data_copy<decvalueT>(copy, nullptr, true);
+
+          ttg::SplitMetadataDescriptor<decvalueT> descr;
+          auto metadata = descr.get_metadata(value);
+          size_t metadata_size = sizeof(metadata);
+          /* pack the metadata */
+          std::memcpy(msg->bytes + pos, &metadata, metadata_size);
+          pos += metadata_size;
+          /* pack the local rank */
+          int rank = world.rank();
+          std::memcpy(msg->bytes + pos, &rank, sizeof(rank));
+          pos += sizeof(rank);
+
+          auto iovecs = descr.get_data(*static_cast<decvalueT *>(copy->device_private));
+
+          int32_t num_iovs = std::distance(std::begin(iovecs), std::end(iovecs));
+          std::memcpy(msg->bytes + pos, &num_iovs, sizeof(num_iovs));
+          pos += sizeof(num_iovs);
+
+          /* TODO: at the moment, the tag argument to parsec_ce.get() is treated as a
+           * raw function pointer instead of a preregistered AM tag, so play that game.
+           * Once this is fixed in PaRSEC we need to use parsec_ttg_rma_tag instead! */
+          parsec_ce_tag_t cbtag = reinterpret_cast<parsec_ce_tag_t>(&detail::get_remote_complete_cb);
+          std::memcpy(msg->bytes + pos, &cbtag, sizeof(cbtag));
+          pos += sizeof(cbtag);
+
+          /**
+           * register the generic iovecs and pack the registration handles
+           * memory layout: [<lreg_size, lreg, release_cb_ptr>, ...]
+           */
+          for (auto &&iov : iovecs) {
+            parsec_ce_mem_reg_handle_t lreg;
+            size_t lreg_size;
+            /* TODO: only register once when we can broadcast the data! */
+            parsec_ce.mem_register(iov.data, PARSEC_MEM_TYPE_NONCONTIGUOUS, iov.num_bytes, parsec_datatype_int8_t,
+                                   iov.num_bytes, &lreg, &lreg_size);
+            auto lreg_ptr = std::shared_ptr<void>{lreg, [](void *ptr) {
+                                                    parsec_ce_mem_reg_handle_t memreg = (parsec_ce_mem_reg_handle_t)ptr;
+                                                    parsec_ce.mem_unregister(&memreg);
+                                                  }};
+            int32_t lreg_size_i = lreg_size;
+            std::memcpy(msg->bytes + pos, &lreg_size_i, sizeof(lreg_size_i));
+            pos += sizeof(lreg_size_i);
+            std::memcpy(msg->bytes + pos, lreg, lreg_size_i);
+            pos += lreg_size_i;
+            /* TODO: can we avoid the extra indirection of going through std::function? */
+            std::function<void(void)> *fn = new std::function<void(void)>([=]() mutable {
+              /* shared_ptr of value and registration captured by value so resetting
+               * them here will eventually release the memory/registration */
+              detail::release_data_copy(copy);
+              lreg_ptr.reset();
+            });
+            std::intptr_t fn_ptr{reinterpret_cast<std::intptr_t>(fn)};
+            std::memcpy(msg->bytes + pos, &fn_ptr, sizeof(fn_ptr));
+            pos += sizeof(fn_ptr);
+          }
+        }
+      }
+      parsec_taskpool_t *tp = world_impl.taskpool();
+      tp->tdm.module->outgoing_message_start(tp, owner, NULL);
+      tp->tdm.module->outgoing_message_pack(tp, owner, NULL, NULL, 0);
+      // std::cout << "Sending AM with " << msg->op_id.num_keys << " keys " << std::endl;
+      parsec_ce.send_am(&parsec_ce, world_impl.parsec_ttg_tag(), owner, static_cast<void *>(msg.get()),
+                        sizeof(detail::msg_header_t) + pos);
+    }
+
+    template <int i, typename Iterator, typename Value>
+    void broadcast_arg_local(Iterator &&begin, Iterator &&end, const Value &value) {
+      parsec_task_t *task_ring = nullptr;
+      detail::ttg_data_copy_t *copy = nullptr;
+      if (nullptr != detail::parsec_ttg_caller) {
+        copy = detail::find_copy_in_task(detail::parsec_ttg_caller, &value);
+      }
+
+      for (auto it = begin; it != end; ++it) {
+        set_arg_local_impl<i>(*it, value, copy, &task_ring);
+      }
+      /* submit all ready tasks at once */
+      if (nullptr != task_ring) {
+        __parsec_schedule(world.impl().execution_stream(), task_ring, 0);
+      }
+    }
+
+    template <std::size_t i, typename Key, typename Value>
+    std::enable_if_t<!ttg::meta::is_void_v<Key> && !std::is_void_v<std::decay_t<Value>> &&
+                         !ttg::has_split_metadata<std::decay_t<Value>>::value,
+                     void>
+    broadcast_arg(const ttg::span<const Key> &keylist, const Value &value) {
+      auto world = ttg_default_execution_context();
+      int rank = world.rank();
+
+      bool have_remote = keylist.end() != std::find_if(keylist.begin(), keylist.end(),
+                                                       [&](const Key &key) { return keymap(key) != rank; });
+
+      if (have_remote) {
+        std::vector<Key> keylist_sorted(keylist.begin(), keylist.end());
+
+        /* Assuming there are no local keys, will be updated while processing remote keys */
+        auto local_begin = keylist_sorted.end();
+        auto local_end = keylist_sorted.end();
+
+        /* sort the input key list by owner and check whether there are remote keys */
+        std::sort(keylist_sorted.begin(), keylist_sorted.end(), [&](const Key &a, const Key &b) mutable {
+          int rank_a = keymap(a);
+          int rank_b = keymap(b);
+          return rank_a < rank_b;
+        });
+
+        using msg_t = detail::msg_t;
+        local_begin = keylist_sorted.end();
+        auto &world_impl = world.impl();
+        std::unique_ptr<msg_t> msg = std::make_unique<msg_t>(get_instance_id(), world_impl.taskpool()->taskpool_id,
+                                                             detail::msg_header_t::MSG_SET_ARG, i);
+
+        parsec_taskpool_t *tp = world_impl.taskpool();
+
+        for (auto it = keylist_sorted.begin(); it < keylist_sorted.end(); /* increment inline */) {
+          auto owner = keymap(*it);
+          if (owner == rank) {
+            /* make sure we don't lose local keys */
+            local_begin = it;
+            local_end =
+                std::find_if_not(++it, keylist_sorted.end(), [&](const Key &key) { return keymap(key) == rank; });
+            it = local_end;
+            continue;
+          }
+
+          /* pack all keys for this owner */
+          int num_keys = 0;
+          uint64_t pos = 0;
+          do {
+            ++num_keys;
+            pos = pack(*it, msg->bytes, pos);
+            ++it;
+          } while (it < keylist_sorted.end() && keymap(*it) == owner);
+          msg->tt_id.num_keys = num_keys;
+
+          /* TODO: use RMA to transfer the value */
+          pos = pack(value, msg->bytes, pos);
+
+          /* Send the message */
+          tp->tdm.module->outgoing_message_start(tp, owner, NULL);
+          tp->tdm.module->outgoing_message_pack(tp, owner, NULL, NULL, 0);
+          parsec_ce.send_am(&parsec_ce, world_impl.parsec_ttg_tag(), owner, static_cast<void *>(msg.get()),
+                            sizeof(detail::msg_header_t) + pos);
+        }
+        /* handle local keys */
+        broadcast_arg_local<i>(local_begin, local_end, value);
+      } else {
+        /* only local keys */
+        broadcast_arg_local<i>(keylist.begin(), keylist.end(), value);
+      }
+    }
+
+    template <std::size_t i, typename Key, typename Value>
+    std::enable_if_t<!ttg::meta::is_void_v<Key> && !std::is_void_v<std::decay_t<Value>> &&
+                         ttg::has_split_metadata<std::decay_t<Value>>::value,
+                     void>
+    splitmd_broadcast_arg(const ttg::span<const Key> &keylist, const Value &value) {
+      using valueT = std::tuple_element_t<i, input_values_full_tuple_type>;
+      auto world = ttg_default_execution_context();
+      int rank = world.rank();
+      bool have_remote = keylist.end() != std::find_if(keylist.begin(), keylist.end(),
+                                                       [&](const Key &key) { return keymap(key) != rank; });
+
+      if (have_remote) {
+        using decvalueT = std::decay_t<Value>;
+
+        /* sort the input key list by owner and check whether there are remote keys */
+        std::vector<Key> keylist_sorted(keylist.begin(), keylist.end());
+        std::sort(keylist_sorted.begin(), keylist_sorted.end(), [&](const Key &a, const Key &b) mutable {
+          int rank_a = keymap(a);
+          int rank_b = keymap(b);
+          return rank_a < rank_b;
+        });
+
+        /* Assuming there are no local keys, will be updated while iterating over the keys */
+        auto local_begin = keylist_sorted.end();
+        auto local_end = keylist_sorted.end();
+
+        ttg::SplitMetadataDescriptor<decvalueT> descr;
+        auto iovs = descr.get_data(*const_cast<decvalueT *>(&value));
+        int32_t num_iovs = std::distance(std::begin(iovs), std::end(iovs));
+        std::vector<std::pair<int32_t, std::shared_ptr<void>>> memregs;
+        memregs.reserve(num_iovs);
+
+        /* register all iovs so the registration can be reused */
+        for (auto &&iov : iovs) {
+          parsec_ce_mem_reg_handle_t lreg;
+          size_t lreg_size;
+          parsec_ce.mem_register(iov.data, PARSEC_MEM_TYPE_NONCONTIGUOUS, iov.num_bytes, parsec_datatype_int8_t,
+                                 iov.num_bytes, &lreg, &lreg_size);
+          /* TODO: use a static function for deregistration here? */
+          memregs.push_back(std::make_pair(static_cast<int32_t>(lreg_size),
+                                           /* TODO: this assumes that parsec_ce_mem_reg_handle_t is void* */
+                                           std::shared_ptr<void>{lreg, [](void *ptr) {
+                                                                   parsec_ce_mem_reg_handle_t memreg =
+                                                                       (parsec_ce_mem_reg_handle_t)ptr;
+                                                                   parsec_ce.mem_unregister(&memreg);
+                                                                 }}));
+        }
+
+        using msg_t = detail::msg_t;
+        auto &world_impl = world.impl();
+        std::unique_ptr<msg_t> msg = std::make_unique<msg_t>(get_instance_id(), world_impl.taskpool()->taskpool_id,
+                                                             detail::msg_header_t::MSG_SET_ARG, i);
+        auto metadata = descr.get_metadata(value);
+        size_t metadata_size = sizeof(metadata);
+
+        detail::ttg_data_copy_t *copy;
+        copy = detail::find_copy_in_task(detail::parsec_ttg_caller, &value);
+        assert(nullptr != copy);
+
+        parsec_taskpool_t *tp = world_impl.taskpool();
+        for (auto it = keylist_sorted.begin(); it < keylist_sorted.end(); /* increment done inline */) {
+          auto owner = keymap(*it);
+          if (owner == rank) {
+            local_begin = it;
+            /* find first non-local key */
+            local_end =
+                std::find_if_not(++it, keylist_sorted.end(), [&](const Key &key) { return keymap(key) == rank; });
+            it = local_end;
+            continue;
+          }
+
+          /* count keys and set it afterwards */
+          uint64_t pos = 0;
+          /* pack all keys for this owner */
+          int num_keys = 0;
+          do {
+            ++num_keys;
+            pos = pack(*it, msg->bytes, pos);
+            ++it;
+          } while (it < keylist_sorted.end() && keymap(*it) == owner);
+          msg->tt_id.num_keys = num_keys;
+
+          /* pack the metadata */
+          std::memcpy(msg->bytes + pos, &metadata, metadata_size);
+          pos += metadata_size;
+          /* pack the local rank */
+          int rank = world.rank();
+          std::memcpy(msg->bytes + pos, &rank, sizeof(rank));
+          pos += sizeof(rank);
+          /* pack the number of iovecs */
+          std::memcpy(msg->bytes + pos, &num_iovs, sizeof(num_iovs));
+          pos += sizeof(num_iovs);
+
+          /* TODO: at the moment, the tag argument to parsec_ce.get() is treated as a
+           * raw function pointer instead of a preregistered AM tag, so play that game.
+           * Once this is fixed in PaRSEC we need to use parsec_ttg_rma_tag instead! */
+          parsec_ce_tag_t cbtag = reinterpret_cast<parsec_ce_tag_t>(&detail::get_remote_complete_cb);
+          std::memcpy(msg->bytes + pos, &cbtag, sizeof(cbtag));
+          pos += sizeof(cbtag);
+
+          /**
+           * pack the registration handles
+           * memory layout: [<lreg_size, lreg, lreg_fn>, ...]
+           */
+          int idx = 0;
+          for (auto &&iov : iovs) {
+            // auto [lreg_size, lreg_ptr] = memregs[idx];
+            int32_t lreg_size;
+            std::shared_ptr<void> lreg_ptr;
+            std::tie(lreg_size, lreg_ptr) = memregs[idx];
+            std::memcpy(msg->bytes + pos, &lreg_size, sizeof(lreg_size));
+            pos += sizeof(lreg_size);
+            std::memcpy(msg->bytes + pos, lreg_ptr.get(), lreg_size);
+            pos += lreg_size;
+            /* create a function that will be invoked upon RMA completion at the target */
+            std::shared_ptr<void> lreg_ptr_v = lreg_ptr;
+            /* mark another reader on the copy */
+            copy = detail::register_data_copy<valueT>(copy, nullptr, true);
+            std::function<void(void)> *fn = new std::function<void(void)>([=]() mutable {
+              /* shared_ptr of value and registration captured by value so resetting
+               * them here will eventually release the memory/registration */
+              detail::release_data_copy(copy);
+              lreg_ptr_v.reset();
+            });
+            std::intptr_t fn_ptr{reinterpret_cast<std::intptr_t>(fn)};
+            std::memcpy(msg->bytes + pos, &fn_ptr, sizeof(fn_ptr));
+            pos += sizeof(fn_ptr);
+            ++idx;
+          }
+          tp->tdm.module->outgoing_message_start(tp, owner, NULL);
+          tp->tdm.module->outgoing_message_pack(tp, owner, NULL, NULL, 0);
+          parsec_ce.send_am(&parsec_ce, world_impl.parsec_ttg_tag(), owner, static_cast<void *>(msg.get()),
+                            sizeof(detail::msg_header_t) + pos);
+        }
+        /* handle local keys */
+        broadcast_arg_local<i>(local_begin, local_end, value);
+      } else {
+        /* handle local keys */
+        broadcast_arg_local<i>(keylist.begin(), keylist.end(), value);
+      }
+    }
+
+    // Used by invoke to set all arguments associated with a task
+    // Is: index sequence of elements in args
+    // Js: index sequence of input terminals to set
+    template <typename Key, typename... Ts, size_t... Is, size_t... Js>
+    std::enable_if_t<ttg::meta::is_none_void_v<Key>, void> set_args(std::index_sequence<Is...>,
+                                                                    std::index_sequence<Js...>, const Key &key,
+                                                                    const std::tuple<Ts...> &args) {
+      static_assert(sizeof...(Js) == sizeof...(Is));
+      constexpr size_t js[] = {Js...};
+      int junk[] = {0, (set_arg<js[Is]>(key, TT::get<Is>(args)), 0)...};
+      junk[0]++;
+    }
+
+    // Used by invoke to set all arguments associated with a task
+    // Is: index sequence of input terminals to set
+    template <typename Key, typename... Ts, size_t... Is>
+    std::enable_if_t<ttg::meta::is_none_void_v<Key>, void> set_args(std::index_sequence<Is...> is, const Key &key,
+                                                                    const std::tuple<Ts...> &args) {
+      set_args(std::index_sequence_for<Ts...>{}, is, key, args);
+    }
+
+    // Used by invoke to set all arguments associated with a task
+    // Is: index sequence of elements in args
+    // Js: index sequence of input terminals to set
+    template <typename Key = keyT, typename... Ts, size_t... Is, size_t... Js>
+    std::enable_if_t<ttg::meta::is_void_v<Key>, void> set_args(std::index_sequence<Is...>, std::index_sequence<Js...>,
+                                                               const std::tuple<Ts...> &args) {
+      static_assert(sizeof...(Js) == sizeof...(Is));
+      constexpr size_t js[] = {Js...};
+      int junk[] = {0, (set_arg<js[Is], void>(TT::get<Is>(args)), 0)...};
+      junk[0]++;
+    }
+
+    // Used by invoke to set all arguments associated with a task
+    // Is: index sequence of input terminals to set
+    template <typename Key = keyT, typename... Ts, size_t... Is>
+    std::enable_if_t<ttg::meta::is_void_v<Key>, void> set_args(std::index_sequence<Is...> is,
+                                                               const std::tuple<Ts...> &args) {
+      set_args(std::index_sequence_for<Ts...>{}, is, args);
+    }
+
+   public:
+    /// sets the default stream size for input \c i
+    /// \param size positive integer that specifies the default stream size
+    template <std::size_t i>
+    void set_static_argstream_size(std::size_t size) {
+      assert(std::get<i>(input_reducers) && "TT::set_argstream_size called on nonstreaming input terminal");
+      assert(size > 0 && "TT::set_static_argstream_size(key,size) called with size=0");
+
+      this->trace(world.rank(), ":", get_name(), ": setting global stream size for terminal ", i);
+
+      // Check if stream is already bounded
+      if (static_stream_goal[i] > 0) {
+        ttg::print_error(world.rank(), ":", get_name(), " : error stream is already bounded : ", i);
+        throw std::runtime_error("TT::set_static_argstream_size called for a bounded stream");
+      }
+
+      static_stream_goal[i] = size;
+    }
+
+    /// sets stream size for input \c i
+    /// \param size positive integer that specifies the stream size
+    /// \param key the task identifier that expects this number of inputs in the streaming terminal
+    template <std::size_t i, typename Key>
+    std::enable_if_t<!ttg::meta::is_void_v<Key>, void> set_argstream_size(const Key &key, std::size_t size) {
+      // preconditions
+      assert(std::get<i>(input_reducers) && "TT::set_argstream_size called on nonstreaming input terminal");
+      assert(size > 0 && "TT::set_argstream_size(key,size) called with size=0");
+
+      // body
+      const auto owner = keymap(key);
+      if (owner != world.rank()) {
+        ttg::trace(world.rank(), ":", get_name(), ":", key, " : forwarding stream size for terminal ", i);
+        using msg_t = detail::msg_t;
+        auto &world_impl = world.impl();
+        uint64_t pos = 0;
+        std::unique_ptr<msg_t> msg = std::make_unique<msg_t>(get_instance_id(), world_impl.taskpool()->taskpool_id,
+                                                             detail::msg_header_t::MSG_SET_ARGSTREAM_SIZE, i, 1);
+        /* pack the key */
+        pos = pack(key, msg->bytes, pos);
+        msg->tt_id.num_keys = 1;
+        pos = pack(size, msg->bytes, pos);
+        parsec_taskpool_t *tp = world_impl.taskpool();
+        tp->tdm.module->outgoing_message_start(tp, owner, NULL);
+        tp->tdm.module->outgoing_message_pack(tp, owner, NULL, NULL, 0);
+        parsec_ce.send_am(&parsec_ce, world_impl.parsec_ttg_tag(), owner, static_cast<void *>(msg.get()),
+                          sizeof(detail::msg_header_t) + pos);
+      } else {
+        ttg::trace(world.rank(), ":", get_name(), ":", key, " : setting stream size to ", size, " for terminal ", i);
+
+        auto hk = reinterpret_cast<parsec_key_t>(&key);
+        task_t *task;
+        parsec_hash_table_lock_bucket(&tasks_table, hk);
+        if (nullptr == (task = (task_t *)parsec_hash_table_nolock_find(&tasks_table, hk))) {
+          task = create_new_task(key);
+          world.impl().increment_created();
+          parsec_hash_table_nolock_insert(&tasks_table, &task->tt_ht_item);
+        }
+
+        // TODO: Unfriendly implementation, cannot check if stream is already bounded
+        // TODO: Unfriendly implementation, cannot check if stream has been finalized already
+
+        // commit changes
+        task->stream[i].goal = size;
+        bool release = (task->stream[i].size == task->stream[i].goal);
+        parsec_hash_table_unlock_bucket(&tasks_table, hk);
+
+        if (release) release_task(task);
+      }
+    }
+
+    /// sets stream size for input \c i
+    /// \param size positive integer that specifies the stream size
+    template <std::size_t i, typename Key = keyT>
+    std::enable_if_t<ttg::meta::is_void_v<Key>, void> set_argstream_size(std::size_t size) {
+      // preconditions
+      assert(std::get<i>(input_reducers) && "TT::set_argstream_size called on nonstreaming input terminal");
+      assert(size > 0 && "TT::set_argstream_size(key,size) called with size=0");
+
+      // body
+      const auto owner = keymap();
+      if (owner != world.rank()) {
+        ttg::trace(world.rank(), ":", get_name(), " : forwarding stream size for terminal ", i);
+        using msg_t = detail::msg_t;
+        auto &world_impl = world.impl();
+        uint64_t pos = 0;
+        std::unique_ptr<msg_t> msg = std::make_unique<msg_t>(get_instance_id(), world_impl.taskpool()->taskpool_id,
+                                                             detail::msg_header_t::MSG_SET_ARGSTREAM_SIZE, i, 1);
+        /* pack the key */
+        msg->tt_id.num_keys = 0;
+        pos = pack(size, msg->bytes, pos);
+        parsec_taskpool_t *tp = world_impl.taskpool();
+        tp->tdm.module->outgoing_message_start(tp, owner, NULL);
+        tp->tdm.module->outgoing_message_pack(tp, owner, NULL, NULL, 0);
+        parsec_ce.send_am(&parsec_ce, world_impl.parsec_ttg_tag(), owner, static_cast<void *>(msg.get()),
+                          sizeof(detail::msg_header_t) + pos);
+      } else {
+        ttg::trace(world.rank(), ":", get_name(), " : setting stream size to ", size, " for terminal ", i);
+
+        parsec_key_t hk = 0;
+        task_t *task;
+        parsec_hash_table_lock_bucket(&tasks_table, hk);
+        if (nullptr == (task = (task_t *)parsec_hash_table_nolock_find(&tasks_table, hk))) {
+          task = create_new_task(ttg::Void{});
+          world.impl().increment_created();
+          parsec_hash_table_nolock_insert(&tasks_table, &task->tt_ht_item);
+        }
+
+        // TODO: Unfriendly implementation, cannot check if stream is already bounded
+        // TODO: Unfriendly implementation, cannot check if stream has been finalized already
+
+        // commit changes
+        task->stream[i].goal = size;
+        bool release = (task->stream[i].size == task->stream[i].goal);
+        parsec_hash_table_unlock_bucket(&tasks_table, hk);
+
+        if (release) release_task(task);
+      }
+    }
+
+    /// finalizes stream for input \c i
+    template <std::size_t i, typename Key>
+    std::enable_if_t<!ttg::meta::is_void_v<Key>, void> finalize_argstream(const Key &key) {
+      // preconditions
+      assert(std::get<i>(input_reducers) && "TT::finalize_argstream called on nonstreaming input terminal");
+
+      // body
+      const auto owner = keymap(key);
+      if (owner != world.rank()) {
+        ttg::trace(world.rank(), ":", get_name(), " : ", key, ": forwarding stream finalize for terminal ", i);
+        using msg_t = detail::msg_t;
+        auto &world_impl = world.impl();
+        uint64_t pos = 0;
+        std::unique_ptr<msg_t> msg = std::make_unique<msg_t>(get_instance_id(), world_impl.taskpool()->taskpool_id,
+                                                             detail::msg_header_t::MSG_FINALIZE_ARGSTREAM_SIZE, i, 1);
+        /* pack the key */
+        pos = pack(key, msg->bytes, pos);
+        msg->tt_id.num_keys = 1;
+        parsec_taskpool_t *tp = world_impl.taskpool();
+        tp->tdm.module->outgoing_message_start(tp, owner, NULL);
+        tp->tdm.module->outgoing_message_pack(tp, owner, NULL, NULL, 0);
+        parsec_ce.send_am(&parsec_ce, world_impl.parsec_ttg_tag(), owner, static_cast<void *>(msg.get()),
+                          sizeof(detail::msg_header_t) + pos);
+      } else {
+        ttg::trace(world.rank(), ":", get_name(), " : ", key, ": finalizing stream for terminal ", i);
+
+        auto hk = reinterpret_cast<parsec_key_t>(&key);
+        task_t *task = nullptr;
+        parsec_hash_table_lock_bucket(&tasks_table, hk);
+        if (nullptr == (task = (task_t *)parsec_hash_table_nolock_find(&tasks_table, hk))) {
+          ttg::print_error(world.rank(), ":", get_name(), ":", key,
+                           " : error finalize called on stream that never received an input data: ", i);
+          throw std::runtime_error("TT::finalize called on stream that never received an input data");
+        }
+
+        // TODO: Unfriendly implementation, cannot check if stream is already bounded
+        // TODO: Unfriendly implementation, cannot check if stream has been finalized already
+
+        // commit changes
+        task->stream[i].size = 1;
+        parsec_hash_table_unlock_bucket(&tasks_table, hk);
+
+        release_task(task);
+      }
+    }
+
+    /// finalizes stream for input \c i
+    template <std::size_t i, bool key_is_void = ttg::meta::is_void_v<keyT>>
+    std::enable_if_t<key_is_void, void> finalize_argstream() {
+      // preconditions
+      assert(std::get<i>(input_reducers) && "TT::finalize_argstream called on nonstreaming input terminal");
+
+      // body
+      const auto owner = keymap();
+      if (owner != world.rank()) {
+        ttg::trace(world.rank(), ":", get_name(), ": forwarding stream finalize for terminal ", i);
+        using msg_t = detail::msg_t;
+        auto &world_impl = world.impl();
+        uint64_t pos = 0;
+        std::unique_ptr<msg_t> msg = std::make_unique<msg_t>(get_instance_id(), world_impl.taskpool()->taskpool_id,
+                                                             detail::msg_header_t::MSG_FINALIZE_ARGSTREAM_SIZE, i, 1);
+        msg->tt_id.num_keys = 0;
+        parsec_taskpool_t *tp = world_impl.taskpool();
+        tp->tdm.module->outgoing_message_start(tp, owner, NULL);
+        tp->tdm.module->outgoing_message_pack(tp, owner, NULL, NULL, 0);
+        parsec_ce.send_am(&parsec_ce, world_impl.parsec_ttg_tag(), owner, static_cast<void *>(msg.get()),
+                          sizeof(detail::msg_header_t) + pos);
+      } else {
+        ttg::trace(world.rank(), ":", get_name(), ": finalizing stream for terminal ", i);
+
+        auto hk = static_cast<parsec_key_t>(0);
+        task_t *task = nullptr;
+        parsec_hash_table_lock_bucket(&tasks_table, hk);
+        if (nullptr == (task = (task_t *)parsec_hash_table_nolock_find(&tasks_table, hk))) {
+          ttg::print_error(world.rank(), ":", get_name(),
+                           " : error finalize called on stream that never received an input data: ", i);
+          throw std::runtime_error("TT::finalize called on stream that never received an input data");
+        }
+
+        // TODO: Unfriendly implementation, cannot check if stream is already bounded
+        // TODO: Unfriendly implementation, cannot check if stream has been finalized already
+
+        // commit changes
+        task->stream[i].size = 1;
+        parsec_hash_table_unlock_bucket(&tasks_table, hk);
+
+        release_task(task);
+      }
+    }
+
+   private:
+    // Copy/assign/move forbidden ... we could make it work using
+    // PIMPL for this base class.  However, this instance of the base
+    // class is tied to a specific instance of a derived class a
+    // pointer to which is captured for invoking derived class
+    // functions.  Thus, not only does the derived class has to be
+    // involved but we would have to do it in a thread safe way
+    // including for possibly already running tasks and remote
+    // references.  This is not worth the effort ... wherever you are
+    // wanting to move/assign an TT you should be using a pointer.
+    TT(const TT &other) = delete;
+    TT &operator=(const TT &other) = delete;
+    TT(TT &&other) = delete;
+    TT &operator=(TT &&other) = delete;
+
+    // Registers the callback for the i'th input terminal
+    template <typename terminalT, std::size_t i>
+    void register_input_callback(terminalT &input) {
+      using valueT = typename terminalT::value_type;
+      if (input.is_pull_terminal) {
+        num_pullins++;
+      }
+      //////////////////////////////////////////////////////////////////
+      // case 1: nonvoid key, nonvoid value
+      //////////////////////////////////////////////////////////////////
+      if constexpr (!ttg::meta::is_void_v<keyT> && !std::is_void_v<valueT>) {
+        auto move_callback = [this](const keyT &key, valueT &&value) {
+          set_arg<i, keyT, valueT>(key, std::forward<valueT>(value));
+        };
+        auto send_callback = [this](const keyT &key, const valueT &value) {
+          set_arg<i, keyT, const valueT &>(key, value);
+        };
+        auto broadcast_callback = [this](const ttg::span<const keyT> &keylist, const valueT &value) {
+          if constexpr (ttg::has_split_metadata<std::decay_t<valueT>>::value) {
+            splitmd_broadcast_arg<i, keyT, valueT>(keylist, value);
+          } else {
+            broadcast_arg<i, keyT, valueT>(keylist, value);
+          }
+        };
+        auto setsize_callback = [this](const keyT &key, std::size_t size) { set_argstream_size<i>(key, size); };
+        auto finalize_callback = [this](const keyT &key) { finalize_argstream<i>(key); };
+        input.set_callback(send_callback, move_callback, broadcast_callback, setsize_callback, finalize_callback);
+      }
+      //////////////////////////////////////////////////////////////////
+      // case 2: nonvoid key, void value, mixed inputs
+      //////////////////////////////////////////////////////////////////
+      else if constexpr (!ttg::meta::is_void_v<keyT> && std::is_void_v<valueT>) {
+        auto send_callback = [this](const keyT &key) { set_arg<i, keyT, ttg::Void>(key, ttg::Void{}); };
+        auto setsize_callback = [this](const keyT &key, std::size_t size) { set_argstream_size<i>(key, size); };
+        auto finalize_callback = [this](const keyT &key) { finalize_argstream<i>(key); };
+        input.set_callback(send_callback, send_callback, {}, setsize_callback, finalize_callback);
+      }
+      //////////////////////////////////////////////////////////////////
+      // case 3: nonvoid key, void value, no inputs
+      // NOTE: subsumed in case 2 above, kept for historical reasons
+      //////////////////////////////////////////////////////////////////
+      //////////////////////////////////////////////////////////////////
+      // case 4: void key, nonvoid value
+      //////////////////////////////////////////////////////////////////
+      else if constexpr (ttg::meta::is_void_v<keyT> && !std::is_void_v<valueT>) {
+        auto move_callback = [this](valueT &&value) { set_arg<i, keyT, valueT>(std::forward<valueT>(value)); };
+        auto send_callback = [this](const valueT &value) { set_arg<i, keyT, const valueT &>(value); };
+        auto setsize_callback = [this](std::size_t size) { set_argstream_size<i>(size); };
+        auto finalize_callback = [this]() { finalize_argstream<i>(); };
+        input.set_callback(send_callback, move_callback, {}, setsize_callback, finalize_callback);
+      }
+      //////////////////////////////////////////////////////////////////
+      // case 5: void key, void value, mixed inputs
+      //////////////////////////////////////////////////////////////////
+      else if constexpr (ttg::meta::is_void_v<keyT> && std::is_void_v<valueT>) {
+        auto send_callback = [this]() { set_arg<i, keyT, ttg::Void>(ttg::Void{}); };
+        auto setsize_callback = [this](std::size_t size) { set_argstream_size<i>(size); };
+        auto finalize_callback = [this]() { finalize_argstream<i>(); };
+        input.set_callback(send_callback, send_callback, {}, setsize_callback, finalize_callback);
+      }
+      //////////////////////////////////////////////////////////////////
+      // case 6: void key, void value, no inputs
+      // NOTE: subsumed in case 5 above, kept for historical reasons
+      //////////////////////////////////////////////////////////////////
+      else
+        abort();
+    }
+
+    template <std::size_t... IS>
+    void register_input_callbacks(std::index_sequence<IS...>) {
+      int junk[] = {
+          0,
+          (register_input_callback<std::tuple_element_t<IS, input_terminals_type>, IS>(std::get<IS>(input_terminals)),
+           0)...};
+      junk[0]++;
+    }
+
+    template <std::size_t... IS, typename inedgesT>
+    void connect_my_inputs_to_incoming_edge_outputs(std::index_sequence<IS...>, inedgesT &inedges) {
+      int junk[] = {0, (std::get<IS>(inedges).set_out(&std::get<IS>(input_terminals)), 0)...};
+      junk[0]++;
+    }
+
+    template <std::size_t... IS, typename outedgesT>
+    void connect_my_outputs_to_outgoing_edge_inputs(std::index_sequence<IS...>, outedgesT &outedges) {
+      int junk[] = {0, (std::get<IS>(outedges).set_in(&std::get<IS>(output_terminals)), 0)...};
+      junk[0]++;
+    }
+
+    template <typename input_terminals_tupleT, std::size_t... IS, typename flowsT>
+    void _initialize_flows(std::index_sequence<IS...>, flowsT &&flows) {
+      int junk[] = {0,
+                    (*(const_cast<std::remove_const_t<decltype(flows[IS]->flow_flags)> *>(&(flows[IS]->flow_flags))) =
+                         (std::is_const_v<std::tuple_element_t<IS, input_terminals_tupleT>> ? PARSEC_FLOW_ACCESS_READ
+                                                                                            : PARSEC_FLOW_ACCESS_RW),
+                     0)...};
+      junk[0]++;
+    }
+
+    template <typename input_terminals_tupleT, typename flowsT>
+    void initialize_flows(flowsT &&flows) {
+      _initialize_flows<input_terminals_tupleT>(
+          std::make_index_sequence<std::tuple_size<input_terminals_tupleT>::value>{}, flows);
+    }
+
+    void fence() override { ttg::default_execution_context().impl().fence(); }
+
+    static int key_equal(parsec_key_t a, parsec_key_t b, void *user_data) {
+      if constexpr (std::is_same_v<keyT, void>) {
+        return 1;
+      } else {
+        keyT &ka = *(reinterpret_cast<keyT *>(a));
+        keyT &kb = *(reinterpret_cast<keyT *>(b));
+        return ka == kb;
+      }
+    }
+
+    static uint64_t key_hash(parsec_key_t k, void *user_data) {
+      constexpr const bool keyT_is_Void = ttg::meta::is_void_v<keyT>;
+      if constexpr (keyT_is_Void || std::is_same_v<keyT, void>) {
+        return 0;
+      } else {
+        keyT &kk = *(reinterpret_cast<keyT *>(k));
+        using ttg::hash;
+        uint64_t hv = hash<decltype(kk)>{}(kk);
+        return hv;
+      }
+    }
+
+    static char *key_print(char *buffer, size_t buffer_size, parsec_key_t k, void *user_data) {
+      if constexpr (std::is_same_v<keyT, void>) {
+        buffer[0] = '\0';
+        return buffer;
+      } else {
+        keyT kk = *(reinterpret_cast<keyT *>(k));
+        std::stringstream iss;
+        iss << kk;
+        memset(buffer, 0, buffer_size);
+        iss.get(buffer, buffer_size);
+        return buffer;
+      }
+    }
+
+    parsec_key_fn_t tasks_hash_fcts = {key_equal, key_print, key_hash};
+
+    static parsec_hook_return_t complete_task_and_release(parsec_execution_stream_t *es, parsec_task_t *t) {
+      parsec_execution_stream_t *safe_es = detail::parsec_ttg_es;
+      detail::parsec_ttg_es = es;
+      auto *task = (detail::parsec_ttg_task_base_t *)t;
+      for (int i = 0; i < task->data_count; i++) {
+        detail::ttg_data_copy_t *copy = static_cast<detail::ttg_data_copy_t *>(task->parsec_task.data[i].data_in);
+        if (nullptr == copy) continue;
+        detail::release_data_copy(copy);
+        task->parsec_task.data[i].data_in = nullptr;
+      }
+      detail::parsec_ttg_es = safe_es;
+      return PARSEC_HOOK_RETURN_DONE;
+    }
+
+   public:
+    template <typename keymapT = ttg::detail::default_keymap<keyT>,
+              typename priomapT = ttg::detail::default_priomap<keyT>>
+    TT(const std::string &name, const std::vector<std::string> &innames, const std::vector<std::string> &outnames,
+       ttg::World world, keymapT &&keymap_ = keymapT(), priomapT &&priomap_ = priomapT())
+        : ttg::TTBase(name, numinedges, numouts)
+        , world(world)
+        // if using default keymap, rebind to the given world
+        , keymap(std::is_same<keymapT, ttg::detail::default_keymap<keyT>>::value
+                     ? decltype(keymap)(ttg::detail::default_keymap<keyT>(world))
+                     : decltype(keymap)(std::forward<keymapT>(keymap_)))
+        , priomap(decltype(keymap)(std::forward<priomapT>(priomap_)))
+        , static_stream_goal() {
+      // Cannot call these in base constructor since terminals not yet constructed
+      if (innames.size() != numinedges) throw std::logic_error("ttg_parsec::TT: #input names != #input terminals");
+      if (outnames.size() != numouts) throw std::logic_error("ttg_parsec::TT: #output names != #output terminals");
+
+      auto &world_impl = world.impl();
+      world_impl.register_op(this);
+
+      if constexpr (numinedges == numins) {
+        register_input_terminals(input_terminals, innames);
+      } else {
+        // create a name for the virtual control input
+        register_input_terminals(input_terminals, std::array<std::string, 1>{std::string("Virtual Control")});
+      }
+      register_output_terminals(output_terminals, outnames);
+
+      register_input_callbacks(std::make_index_sequence<numinedges>{});
+      int i;
+
+      memset(&self, 0, sizeof(parsec_task_class_t));
+
+      self.name = get_name().c_str();
+      self.task_class_id = get_instance_id();
+      self.nb_parameters = 0;
+      self.nb_locals = 0;
+      self.nb_flows = numflows;
+
+      //    function_id_to_instance[self.task_class_id] = this;
+
+      if constexpr (derived_has_cuda_op()) {
+        self.incarnations = (__parsec_chore_t *)malloc(3 * sizeof(__parsec_chore_t));
+        ((__parsec_chore_t *)self.incarnations)[0].type = PARSEC_DEV_CUDA;
+        ((__parsec_chore_t *)self.incarnations)[0].evaluate = NULL;
+        ((__parsec_chore_t *)self.incarnations)[0].hook = detail::hook_cuda;
+        ((__parsec_chore_t *)self.incarnations)[1].type = PARSEC_DEV_CPU;
+        ((__parsec_chore_t *)self.incarnations)[1].evaluate = NULL;
+        ((__parsec_chore_t *)self.incarnations)[1].hook = detail::hook;
+        ((__parsec_chore_t *)self.incarnations)[2].type = PARSEC_DEV_NONE;
+        ((__parsec_chore_t *)self.incarnations)[2].evaluate = NULL;
+        ((__parsec_chore_t *)self.incarnations)[2].hook = NULL;
+      } else {
+        self.incarnations = (__parsec_chore_t *)malloc(2 * sizeof(__parsec_chore_t));
+        ((__parsec_chore_t *)self.incarnations)[0].type = PARSEC_DEV_CPU;
+        ((__parsec_chore_t *)self.incarnations)[0].evaluate = NULL;
+        ((__parsec_chore_t *)self.incarnations)[0].hook = detail::hook;
+        ((__parsec_chore_t *)self.incarnations)[1].type = PARSEC_DEV_NONE;
+        ((__parsec_chore_t *)self.incarnations)[1].evaluate = NULL;
+        ((__parsec_chore_t *)self.incarnations)[1].hook = NULL;
+      }
+
+      self.release_task = &parsec_release_task_to_mempool_update_nbtasks;
+      self.complete_execution = complete_task_and_release;
+
+      for (i = 0; i < numins; i++) {
+        parsec_flow_t *flow = new parsec_flow_t;
+        flow->name = strdup((std::string("flow in") + std::to_string(i)).c_str());
+        flow->sym_type = PARSEC_SYM_INOUT;
+        // see initialize_flows below
+        // flow->flow_flags = PARSEC_FLOW_ACCESS_RW;
+        flow->dep_in[0] = NULL;
+        flow->dep_out[0] = NULL;
+        flow->flow_index = i;
+        flow->flow_datatype_mask = (1 << i);
+        *((parsec_flow_t **)&(self.in[i])) = flow;
+      }
+      *((parsec_flow_t **)&(self.in[i])) = NULL;
+      initialize_flows<input_terminals_type>(self.in);
+
+      for (i = 0; i < numouts; i++) {
+        parsec_flow_t *flow = new parsec_flow_t;
+        flow->name = strdup((std::string("flow out") + std::to_string(i)).c_str());
+        flow->sym_type = PARSEC_SYM_INOUT;
+        flow->flow_flags = PARSEC_FLOW_ACCESS_READ;  // does PaRSEC use this???
+        flow->dep_in[0] = NULL;
+        flow->dep_out[0] = NULL;
+        flow->flow_index = i;
+        flow->flow_datatype_mask = (1 << i);
+        *((parsec_flow_t **)&(self.out[i])) = flow;
+      }
+      *((parsec_flow_t **)&(self.out[i])) = NULL;
+
+      self.flags = 0;
+      self.dependencies_goal = numins; /* (~(uint32_t)0) >> (32 - numins); */
+
+      int nbthreads = 0;
+      auto *context = world_impl.context();
+      for (int i = 0; i < context->nb_vp; i++) {
+        nbthreads += context->virtual_processes[i]->nb_cores;
+      }
+
+      parsec_mempool_construct(&mempools, PARSEC_OBJ_CLASS(parsec_task_t), sizeof(task_t),
+                               offsetof(parsec_task_t, mempool_owner), nbthreads);
+
+      parsec_hash_table_init(&tasks_table, offsetof(detail::parsec_ttg_task_base_t, tt_ht_item), 8, tasks_hash_fcts,
+                             NULL);
+    }
+
+    template <typename keymapT = ttg::detail::default_keymap<keyT>,
+              typename priomapT = ttg::detail::default_priomap<keyT>>
+    TT(const std::string &name, const std::vector<std::string> &innames, const std::vector<std::string> &outnames,
+       keymapT &&keymap = keymapT(ttg::default_execution_context()), priomapT &&priomap = priomapT())
+        : TT(name, innames, outnames, ttg::default_execution_context(), std::forward<keymapT>(keymap),
+             std::forward<priomapT>(priomap)) {}
+
+    template <typename keymapT = ttg::detail::default_keymap<keyT>,
+              typename priomapT = ttg::detail::default_priomap<keyT>>
+    TT(const input_edges_type &inedges, const output_edges_type &outedges, const std::string &name,
+       const std::vector<std::string> &innames, const std::vector<std::string> &outnames, ttg::World world,
+       keymapT &&keymap_ = keymapT(), priomapT &&priomap = priomapT())
+        : TT(name, innames, outnames, world, std::forward<keymapT>(keymap_), std::forward<priomapT>(priomap)) {
+      connect_my_inputs_to_incoming_edge_outputs(std::make_index_sequence<numinedges>{}, inedges);
+      connect_my_outputs_to_outgoing_edge_inputs(std::make_index_sequence<numouts>{}, outedges);
+      //DO NOT MOVE THIS - information about the number of pull terminals is only available after connecting the edges.
+      if constexpr (numinedges > 0) {
+        register_input_callbacks(std::make_index_sequence<numinedges>{});
+      }
+    }
+    template <typename keymapT = ttg::detail::default_keymap<keyT>,
+              typename priomapT = ttg::detail::default_priomap<keyT>>
+    TT(const input_edges_type &inedges, const output_edges_type &outedges, const std::string &name,
+       const std::vector<std::string> &innames, const std::vector<std::string> &outnames,
+       keymapT &&keymap = keymapT(ttg::default_execution_context()), priomapT &&priomap = priomapT())
+        : TT(inedges, outedges, name, innames, outnames, ttg::default_execution_context(),
+             std::forward<keymapT>(keymap), std::forward<priomapT>(priomap)) {}
+
+    // Destructor checks for unexecuted tasks
+    virtual ~TT() { release(); }
+
+    static void ht_iter_cb(void *item, void *cb_data) {
+      task_t *task = (task_t *)item;
+      ttT *op = (ttT *)cb_data;
+      if constexpr (!ttg::meta::is_void_v<keyT>) {
+        std::cout << "Left over task " << op->get_name() << " " << task->key << std::endl;
+      } else {
+        std::cout << "Left over task " << op->get_name() << std::endl;
+      }
+    }
+
+    virtual void release() override { do_release(); }
+
+    void do_release() {
+      if (!alive) {
+        return;
+      }
+      alive = false;
+      /* print all outstanding tasks */
+      parsec_hash_table_for_all(&tasks_table, ht_iter_cb, this);
+      parsec_hash_table_fini(&tasks_table);
+      parsec_mempool_destruct(&mempools);
+      // uintptr_t addr = (uintptr_t)self.incarnations;
+      // free((void *)addr);
+      free((__parsec_chore_t *)self.incarnations);
+      for (int i = 0; i < numflows; i++) {
+        if (NULL != self.in[i]) {
+          free(self.in[i]->name);
+          delete self.in[i];
+        }
+        if (NULL != self.out[i]) {
+          free(self.out[i]->name);
+          delete self.out[i];
+        }
+      }
+      world.impl().deregister_op(this);
+    }
+
+    static constexpr const ttg::Runtime runtime = ttg::Runtime::PaRSEC;
+
+    /// define the reducer function to be called when additional inputs are
+    /// received on a streaming terminal
+    ///   @tparam <i> the index of the input terminal that is used as a streaming terminal
+    ///   @param[in] reducer: a function of prototype (input_type<i> &a, const input_type<i> &b)
+    ///                       that function should aggregate b into a
+    template <std::size_t i, typename Reducer>
+    void set_input_reducer(Reducer &&reducer) {
+      ttg::trace(world.rank(), ":", get_name(), " : setting reducer for terminal ", i);
+      std::get<i>(input_reducers) = reducer;
+    }
+
+    /// define the reducer function to be called when additional inputs are
+    /// received on a streaming terminal
+    ///   @tparam <i> the index of the input terminal that is used as a streaming terminal
+    ///   @param[in] reducer: a function of prototype (input_type<i> &a, const input_type<i> &b)
+    ///                       that function should aggregate b into a
+    ///   @param[in] size: the default number of inputs that are received in this streaming terminal,
+    ///                    for each task
+    template <std::size_t i, typename Reducer>
+    void set_input_reducer(Reducer &&reducer, std::size_t size) {
+      set_input_reducer<i>(std::forward<Reducer>(reducer));
+      set_static_argstream_size<i>(size);
+    }
+
+    // Returns reference to input terminal i to facilitate connection --- terminal
+    // cannot be copied, moved or assigned
+    template <std::size_t i>
+    std::tuple_element_t<i, input_terminals_type> *in() {
+      return &std::get<i>(input_terminals);
+    }
+
+    // Returns reference to output terminal for purpose of connection --- terminal
+    // cannot be copied, moved or assigned
+    template <std::size_t i>
+    std::tuple_element_t<i, output_terminalsT> *out() {
+      return &std::get<i>(output_terminals);
+    }
+
+    // Manual injection of a task with all input arguments specified as a tuple
+    template <typename Key = keyT>
+    std::enable_if_t<!ttg::meta::is_void_v<Key> && !ttg::meta::is_empty_tuple_v<input_values_tuple_type>, void> invoke(
+        const Key &key, const input_values_tuple_type &args) {
+      TTG_OP_ASSERT_EXECUTABLE();
+      /* trigger non-void inputs */
+      set_args(ttg::meta::nonvoid_index_seq<actual_input_tuple_type>{}, key, args);
+      /* trigger void inputs */
+      using void_index_seq = ttg::meta::void_index_seq<actual_input_tuple_type>;
+      set_args(void_index_seq{}, key, ttg::detail::make_void_tuple<void_index_seq::size()>());
+    }
+
+    // Manual injection of a key-free task and all input arguments specified as a tuple
+    template <typename Key = keyT>
+    std::enable_if_t<ttg::meta::is_void_v<Key> && !ttg::meta::is_empty_tuple_v<input_values_tuple_type>, void> invoke(
+        const input_values_tuple_type &args) {
+      TTG_OP_ASSERT_EXECUTABLE();
+      /* trigger non-void inputs */
+      set_args(ttg::meta::nonvoid_index_seq<actual_input_tuple_type>{}, args);
+      /* trigger void inputs */
+      using void_index_seq = ttg::meta::void_index_seq<actual_input_tuple_type>;
+      set_args(void_index_seq{}, ttg::detail::make_void_tuple<void_index_seq::size()>());
+    }
+
+    // Manual injection of a task that has no arguments
+    template <typename Key = keyT>
+    std::enable_if_t<!ttg::meta::is_void_v<Key> && ttg::meta::is_empty_tuple_v<input_values_tuple_type>, void> invoke(
+        const Key &key) {
+      TTG_OP_ASSERT_EXECUTABLE();
+      /* trigger void inputs */
+      using void_index_seq = ttg::meta::void_index_seq<actual_input_tuple_type>;
+      set_args(void_index_seq{}, key, ttg::detail::make_void_tuple<void_index_seq::size()>());
+    }
+
+    // Manual injection of a task that has no key or arguments
+    template <typename Key = keyT>
+    std::enable_if_t<ttg::meta::is_void_v<Key> && ttg::meta::is_empty_tuple_v<input_values_tuple_type>, void> invoke() {
+      TTG_OP_ASSERT_EXECUTABLE();
+      /* trigger void inputs */
+      using void_index_seq = ttg::meta::void_index_seq<actual_input_tuple_type>;
+      set_args(void_index_seq{}, ttg::detail::make_void_tuple<void_index_seq::size()>());
+    }
+
+    // overrides TTBase::invoke()
+    void invoke() override {
+      if constexpr (ttg::meta::is_void_v<keyT> && ttg::meta::is_empty_tuple_v<input_values_tuple_type>)
+        invoke<keyT>();
+      else
+        TTBase::invoke();
+    }
+
+   public:
+    void make_executable() override {
+      register_static_op_function();
+      ttg::TTBase::make_executable();
+    }
+
+    /// keymap accessor
+    /// @return the keymap
+    const decltype(keymap) &get_keymap() const { return keymap; }
+
+    /// keymap setter
+    template <typename Keymap>
+    void set_keymap(Keymap &&km) {
+      keymap = km;
+    }
+
+    /// priority map accessor
+    /// @return the priority map
+    const decltype(priomap) &get_priomap() const { return priomap; }
+
+    /// priomap setter
+    /// @arg pm a function that maps a key to an integral priority value.
+    template <typename Priomap>
+    void set_priomap(Priomap &&pm) {
+      priomap = pm;
+    }
+
+    // Register the static_op function to associate it to instance_id
+    void register_static_op_function(void) {
+      int rank;
+      MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+      ttg::trace("ttg_parsec(", rank, ") Inserting into static_id_to_op_map at ", get_instance_id());
+      detail::static_set_arg_fct_call_t call = std::make_pair(&TT::static_set_arg, this);
+      auto &world_impl = world.impl();
+      detail::static_map_mutex.lock();
+      detail::static_id_to_op_map.insert(std::make_pair(get_instance_id(), call));
+      if (detail::delayed_unpack_actions.count(get_instance_id()) > 0) {
+        auto tp = world_impl.taskpool();
+
+        ttg::trace("ttg_parsec(", rank, ") There are ", detail::delayed_unpack_actions.count(get_instance_id()),
+                   " messages delayed with op_id ", get_instance_id());
+
+        auto se = detail::delayed_unpack_actions.equal_range(get_instance_id());
+        std::vector<detail::static_set_arg_fct_arg_t> tmp;
+        for (auto it = se.first; it != se.second;) {
+          assert(it->first == get_instance_id());
+          tmp.push_back(it->second);
+          it = detail::delayed_unpack_actions.erase(it);
+        }
+        detail::static_map_mutex.unlock();
+
+        for (auto it : tmp) {
+          ttg::print("ttg_parsec(", rank, ") Unpacking delayed message (", ", ", get_instance_id(), ", ",
+                     std::get<1>(it), ", ", std::get<2>(it), ")");
+          int rc = detail::static_unpack_msg(&parsec_ce, world_impl.parsec_ttg_tag(), std::get<1>(it), std::get<2>(it),
+                                             std::get<0>(it), NULL);
+          assert(rc == 0);
+          free(std::get<1>(it));
+        }
+
+        tmp.clear();
+      } else {
+        detail::static_map_mutex.unlock();
+      }
+    }
+  };
+} // namespace ttg_parsec
+
+#endif // TTG_PARSEC_TT_H
