+++ ttg/ttg/parsec/world.h
@@ -1,0 +1,179 @@
+#ifndef TTG_PARSEC_WORLD_H
+#define TTG_PARSEC_WORLD_H
+
+#include <mpi.h>
+#include <parsec.h>
+#include <parsec/execution_stream.h>
+
+#include "ttg/base/world.h"
+#include "ttg/edge.h"
+#include "ttg/parsec/fwd.h"
+
+#include "ttg/parsec/vars.h"
+
+namespace ttg_parsec {
+
+  class WorldImpl : public ttg::base::WorldImplBase {
+    static constexpr const int _PARSEC_TTG_TAG = 10;      // This TAG should be 'allocated' at the PaRSEC level
+    static constexpr const int _PARSEC_TTG_RMA_TAG = 11;  // This TAG should be 'allocated' at the PaRSEC level
+
+    ttg::Edge<> m_ctl_edge;
+
+    int query_comm_size() {
+      int comm_size;
+      MPI_Comm_size(MPI_COMM_WORLD, &comm_size);
+      return comm_size;
+    }
+
+    int query_comm_rank() {
+      int comm_rank;
+      MPI_Comm_rank(MPI_COMM_WORLD, &comm_rank);
+      return comm_rank;
+    }
+
+   public:
+    static constexpr const int PARSEC_TTG_MAX_AM_SIZE = 1024 * 1024;
+    WorldImpl(int *argc, char **argv[], int ncores) : WorldImplBase(query_comm_size(), query_comm_rank()) {
+      ttg::detail::register_world(*this);
+      ctx = parsec_init(ncores, argc, argv);
+      es = ctx->virtual_processes[0]->execution_streams[0];
+
+      parsec_ce.tag_register(_PARSEC_TTG_TAG, &detail::static_unpack_msg, this, PARSEC_TTG_MAX_AM_SIZE);
+      parsec_ce.tag_register(_PARSEC_TTG_RMA_TAG, &detail::get_remote_complete_cb, this, 128);
+
+      create_tpool();
+    }
+
+    void create_tpool() {
+      assert(nullptr == tpool);
+      tpool = (parsec_taskpool_t *)calloc(1, sizeof(parsec_taskpool_t));
+      tpool->taskpool_id = -1;
+      tpool->update_nb_runtime_task = parsec_add_fetch_runtime_task;
+      tpool->taskpool_type = PARSEC_TASKPOOL_TYPE_TTG;
+      parsec_taskpool_reserve_id(tpool);
+
+#ifdef TTG_USE_USER_TERMDET
+      parsec_termdet_open_module(tpool, "user_trigger");
+#else   // TTG_USE_USER_TERMDET
+      parsec_termdet_open_dyn_module(tpool);
+#endif  // TTG_USE_USER_TERMDET
+      tpool->tdm.module->monitor_taskpool(tpool, parsec_taskpool_termination_detected);
+      // In TTG, we use the pending actions to denote that the
+      // taskpool is not ready, i.e. some local tasks could still
+      // be added by the main thread. It should then be initialized
+      // to 0, execute will set it to 1 and mark the tpool as ready,
+      // and the fence() will decrease it back to 0.
+      tpool->tdm.module->taskpool_set_nb_pa(tpool, 0);
+      parsec_taskpool_enable(tpool, NULL, NULL, es, size() > 1);
+
+      // Termination detection in PaRSEC requires to synchronize the
+      // taskpool enabling, to avoid a race condition that would keep
+      // termination detection-related messages in a waiting queue
+      // forever
+      MPI_Barrier(comm());
+
+      parsec_taskpool_started = false;
+    }
+
+    /* Deleted copy ctor */
+    WorldImpl(const WorldImpl &other) = delete;
+
+    /* Deleted move ctor */
+    WorldImpl(WorldImpl &&other) = delete;
+
+    /* Deleted copy assignment */
+    WorldImpl &operator=(const WorldImpl &other) = delete;
+
+    /* Deleted move assignment */
+    WorldImpl &operator=(WorldImpl &&other) = delete;
+
+    ~WorldImpl() { destroy(); }
+
+    constexpr int parsec_ttg_tag() const { return _PARSEC_TTG_TAG; }
+    constexpr int parsec_ttg_rma_tag() const { return _PARSEC_TTG_RMA_TAG; }
+
+    MPI_Comm comm() const { return MPI_COMM_WORLD; }
+
+    virtual void execute() override {
+      parsec_enqueue(ctx, tpool);
+      tpool->tdm.module->taskpool_addto_nb_pa(tpool, 1);
+      tpool->tdm.module->taskpool_ready(tpool);
+      int ret = parsec_context_start(ctx);
+      parsec_taskpool_started = true;
+      if (ret != 0) throw std::runtime_error("TTG: parsec_context_start failed");
+    }
+
+    void destroy_tpool() {
+      parsec_taskpool_free(tpool);
+      tpool = nullptr;
+    }
+
+    virtual void destroy() override {
+      if (is_valid()) {
+        if (parsec_taskpool_started) {
+          // We are locally ready (i.e. we won't add new tasks)
+          tpool->tdm.module->taskpool_addto_nb_pa(tpool, -1);
+          ttg::trace("ttg_parsec(", this->rank(), "): final waiting for completion");
+          parsec_context_wait(ctx);
+        }
+        release_ops();
+        ttg::detail::deregister_world(*this);
+        destroy_tpool();
+        parsec_ce.tag_unregister(_PARSEC_TTG_TAG);
+        parsec_ce.tag_unregister(_PARSEC_TTG_RMA_TAG);
+        parsec_fini(&ctx);
+        mark_invalid();
+      }
+    }
+
+    ttg::Edge<> &ctl_edge() { return m_ctl_edge; }
+
+    const ttg::Edge<> &ctl_edge() const { return m_ctl_edge; }
+
+    auto *context() { return ctx; }
+    auto *execution_stream() { return detail::parsec_ttg_es == nullptr ? es : detail::parsec_ttg_es; }
+    auto *taskpool() { return tpool; }
+
+    void increment_created() { taskpool()->tdm.module->taskpool_addto_nb_tasks(taskpool(), 1); }
+
+    void increment_inflight_msg() { taskpool()->tdm.module->taskpool_addto_nb_pa(taskpool(), 1); }
+    void decrement_inflight_msg() { taskpool()->tdm.module->taskpool_addto_nb_pa(taskpool(), -1); }
+
+    virtual void final_task() override {
+#ifdef TTG_USE_USER_TERMDET
+      taskpool()->tdm.module->taskpool_set_nb_tasks(taskpool(), 0);
+#endif  // TTG_USE_USER_TERMDET
+    }
+
+   protected:
+    virtual void fence_impl(void) override {
+      int rank = this->rank();
+      if (!parsec_taskpool_started) {
+        ttg::trace("ttg_parsec::(", rank, "): parsec taskpool has not been started, fence is a simple MPI_Barrier");
+        MPI_Barrier(comm());
+        return;
+      }
+      ttg::trace("ttg_parsec::(", rank, "): parsec taskpool is ready for completion");
+      // We are locally ready (i.e. we won't add new tasks)
+      tpool->tdm.module->taskpool_addto_nb_pa(tpool, -1);
+      ttg::trace("ttg_parsec(", rank, "): waiting for completion");
+      parsec_context_wait(ctx);
+
+      // We need the synchronization between the end of the context and the restart of the taskpool
+      // until we use parsec_taskpool_wait and implement an epoch in the PaRSEC taskpool
+      // see Issue #118 (TTG)
+      MPI_Barrier(comm());
+
+      destroy_tpool();
+      create_tpool();
+      execute();
+    }
+
+   private:
+    parsec_context_t *ctx = nullptr;
+    parsec_execution_stream_t *es = nullptr;
+    parsec_taskpool_t *tpool = nullptr;
+    bool parsec_taskpool_started = false;
+  };
+} // namespace ttg_parsec
+#endif // TTG_PARSEC_WORLD_H
